Replaces prior TSP example with TSPLIB instance including costs
